@@ -1,6 +1,6 @@
 module istio.io/istio/tests/taaa/test-artifact
 
-go 1.16
+go 1.19
 
 replace istio.io/istio/prow/asm/tester => ../../../prow/asm/tester
 
