@@ -1059,8 +1059,10 @@ type EgressGatewayConfig struct {
 	// your pod is eligible to be scheduled based on labels on pods that are
 	// already running on the node rather than based on labels on nodes.
 	// There are currently two types of anti-affinity:
-	//    "requiredDuringSchedulingIgnoredDuringExecution"
-	//    "preferredDuringSchedulingIgnoredDuringExecution"
+	//
+	//	"requiredDuringSchedulingIgnoredDuringExecution"
+	//	"preferredDuringSchedulingIgnoredDuringExecution"
+	//
 	// which denote “hard” vs. “soft” requirements, you can define your values
 	// in "podAntiAffinityLabelSelector" and "podAntiAffinityTermLabelSelector"
 	// correspondingly.
@@ -1068,13 +1070,13 @@ type EgressGatewayConfig struct {
 	//
 	// Examples:
 	// podAntiAffinityLabelSelector:
-	//  - key: security
-	//    operator: In
-	//    values: S1,S2
-	//    topologyKey: "kubernetes.io/hostname"
-	//  This pod anti-affinity rule says that the pod requires not to be scheduled
-	//  onto a node if that node is already running a pod with label having key
-	//  “security” and value “S1”.
+	//   - key: security
+	//     operator: In
+	//     values: S1,S2
+	//     topologyKey: "kubernetes.io/hostname"
+	//     This pod anti-affinity rule says that the pod requires not to be scheduled
+	//     onto a node if that node is already running a pod with label having key
+	//     “security” and value “S1”.
 	//
 	// Deprecated: Do not use.
 	PodAntiAffinityLabelSelector []*structpb.Struct `protobuf:"bytes,12,rep,name=podAntiAffinityLabelSelector,proto3" json:"podAntiAffinityLabelSelector,omitempty"`
@@ -1415,10 +1417,11 @@ type GlobalConfig struct {
 	unknownFields protoimpl.UnknownFields
 
 	// Specifies pod scheduling arch(amd64, ppc64le, s390x, arm64) and weight as follows:
-	//   0 - Never scheduled
-	//   1 - Least preferred
-	//   2 - No preference
-	//   3 - Most preferred
+	//
+	//	0 - Never scheduled
+	//	1 - Least preferred
+	//	2 - No preference
+	//	3 - Most preferred
 	//
 	// Deprecated: replaced by the affinity k8s settings which allows architecture nodeAffinity configuration of this behavior.
 	//
@@ -1474,19 +1477,19 @@ type GlobalConfig struct {
 	// it still need to be configured manually).
 	//
 	// meshNetworks:
-	//   network1:
-	//     endpoints:
-	//     - fromCidr: "192.168.0.1/24"
-	//     gateways:
-	//     - address: 1.1.1.1
-	//       port: 80
-	//   network2:
-	//     endpoints:
-	//     - fromRegistry: reg1
-	//     gateways:
-	//     - registryServiceName: istio-ingressgateway.istio-system.svc.cluster.local
-	//       port: 443
 	//
+	//	network1:
+	//	  endpoints:
+	//	  - fromCidr: "192.168.0.1/24"
+	//	  gateways:
+	//	  - address: 1.1.1.1
+	//	    port: 80
+	//	network2:
+	//	  endpoints:
+	//	  - fromRegistry: reg1
+	//	  gateways:
+	//	  - registryServiceName: istio-ingressgateway.istio-system.svc.cluster.local
+	//	    port: 443
 	MeshNetworks *structpb.Struct `protobuf:"bytes,19,opt,name=meshNetworks,proto3" json:"meshNetworks,omitempty"`
 	// Specifies the Configuration for Istio mesh across multiple clusters through Istio gateways.
 	MultiCluster *MultiClusterConfig `protobuf:"bytes,22,opt,name=multiCluster,proto3" json:"multiCluster,omitempty"`
@@ -2727,8 +2730,9 @@ type PilotConfig struct {
 	//
 	// Examples:
 	// env:
-	//   ENV_VAR_1: value1
-	//   ENV_VAR_2: value2
+	//
+	//	ENV_VAR_1: value1
+	//	ENV_VAR_2: value2
 	Env *structpb.Struct `protobuf:"bytes,21,opt,name=env,proto3" json:"env,omitempty"`
 	// K8s rolling update strategy
 	//
@@ -2738,8 +2742,6 @@ type PilotConfig struct {
 	//
 	// Deprecated: Do not use.
 	RollingMaxUnavailable *IntOrString `protobuf:"bytes,25,opt,name=rollingMaxUnavailable,proto3" json:"rollingMaxUnavailable,omitempty"`
-	//
-	//
 	// Deprecated: Do not use.
 	Tolerations []*structpb.Struct `protobuf:"bytes,26,rep,name=tolerations,proto3" json:"tolerations,omitempty"`
 	// if protocol sniffing is enabled for outbound
@@ -4261,7 +4263,7 @@ type SidecarInjectorConfig struct {
 	NeverInjectSelector []*structpb.Struct `protobuf:"bytes,11,rep,name=neverInjectSelector,proto3" json:"neverInjectSelector,omitempty"`
 	// See NeverInjectSelector.
 	AlwaysInjectSelector []*structpb.Struct `protobuf:"bytes,12,rep,name=alwaysInjectSelector,proto3" json:"alwaysInjectSelector,omitempty"`
-	//  If true, webhook or istioctl injector will rewrite PodSpec for liveness health check to redirect request to sidecar. This makes liveness check work even when mTLS is enabled.
+	// If true, webhook or istioctl injector will rewrite PodSpec for liveness health check to redirect request to sidecar. This makes liveness check work even when mTLS is enabled.
 	RewriteAppHTTPProbe *wrapperspb.BoolValue `protobuf:"bytes,16,opt,name=rewriteAppHTTPProbe,proto3" json:"rewriteAppHTTPProbe,omitempty"`
 	// injectedAnnotations are additional annotations that will be added to the pod spec after injection
 	// This is primarily to support PSP annotations.
@@ -4273,10 +4275,11 @@ type SidecarInjectorConfig struct {
 	// Templates defines a set of custom injection templates that can be used. For example, defining:
 	//
 	// templates:
-	//   hello: |
-	//     metadata:
-	//       labels:
-	//         hello: world
+	//
+	//	hello: |
+	//	  metadata:
+	//	    labels:
+	//	      hello: world
 	//
 	// Then starting a pod with the `inject.istio.io/templates: hello` annotation, will result in the pod
 	// being injected with the hello=world labels.
